Refactored test suite
