+++ spacescout_labstats/test/resources/uuid_test.json
@@ -1,5 +1,5 @@
 {
-    "good_uuids": ["7d115cc0-7c3d-448f-a89e-b4953f666949",
+    "correct_uuids": ["7d115cc0-7c3d-448f-a89e-b4953f666949",
         "397557e3-cd05-4b80-862a-6737ccbf1e51",
         "2eca9f78-5c87-4a4f-92df-26c950e419e9",
         "67705b87-1598-4569-8b88-5b4a66976bc5",
