+++ spacescout_labstats/test/test_run_labstats_daemon.py
@@ -1,0 +1,7 @@
+"""
+This file contains tests for run_labstats_daemon.
+"""
+from django.test import TestCase
+
+
+class LabstatsDaemonTest(TestCase):
