+++ spacescout_labstats/test/test_seattle_labstats.py
@@ -1,0 +1,7 @@
+"""
+This file contains tests for seattle_labstats.py.
+"""
+from django.test import TestCase
+
+
+class LabstatsDaemonTest(TestCase):
