+++ spacescout_labstats/tests.py
@@ -1,0 +1,1 @@
+from test import tests, test_online_labstats, test_seattle_labstats
